Update README.md
Hey i am building a website of clothes organisation life hack my website comprises of a menu item with support pages. The index page has a banner with a static image and text, it further comprises of support pages and a video to showcase what the website is about. The pages are inter-linked together to give an appealing user experience while browsing the webpages.